+++ docs/guides/changelog.md
@@ -26,6 +26,7 @@ description: |-
 - Add `meraki_appliance_rf_profile` resource and data source
 - Add `meraki_appliance_rf_profiles` data source
 - BREAKING CHANGE: Rename `per_ssid_settingsXX_*` attributes of `meraki_wireless_rf_profile` resource and data sources to `per_ssid_settings_XX_*`
+- Add `meraki_appliance_sdwan_internet_policies` resource
 
 ## 0.1.1